New Background and torch drain
New background added and also torch drain added when the player has their torch active. This gives a use for the new battery item.

+++ Assets/Scripts/Player/BatteryBar.cs
@@ -8,7 +8,7 @@ public class BatteryBar : MonoBehaviour
 
     public Slider slider;
 
-    public void SetBattery(int battery)
+    public void SetBattery(float battery)
     {
         slider.value = battery;
     }
